+++ NEWS.md
@@ -1,5 +1,7 @@
 # latexdiffr (development version)
 
+* Added a `compile` option to (not) compile the diff.tex file to pdf
+
 # latexdiffr 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.